Move DictionaryPayload into a source file.
